refactor(cli): promote ANSI primitives to src/core/ansi.ts
src/cli/tui/cells.ts was importing stripAnsi/truncate from src/hud/colors.ts
(a feature module), and this PR added inverse() to hud/colors with its only
consumer in flags-view/render.ts. The dependency arrow pointed generic-CLI →
feature, which is the wrong direction under ADR-013 (src/core/ = agent-neutral
logic). It also caused composeScripts to carry the transitive dist/hud/ graph
into the hooks runtime unnecessarily.

Promote all ANSI primitives to src/core/ansi.ts (neutral home), then:
- src/cli/tui/cells.ts       imports stripAnsi, truncate from core/ansi
- src/cli/flags-view/render.ts  imports from core/ansi
- src/cli/agents-view/render.ts imports from core/ansi
- src/hud/colors.ts becomes a re-export barrel (export * from ../core/ansi)
  so all existing HUD component call sites (hud/components/*.ts, hud/render.ts,
  src/cli/commands/agents.ts, src/cli/commands/proxy.ts) are untouched.

Byte-equivalent behaviour: functions are identical implementations.
applies ADR-013; applies PF-017 corollary (one shared definition).
